clean up coranking and density metrics
