+++ src/dimensionality_reduction/metrics/coranking/library.bib
@@ -1,0 +1,62 @@
+
+@misc{lueks2011evaluate,
+  doi = {10.48550/ARXIV.1110.3917},
+  url = {https://arxiv.org/abs/1110.3917},
+  author = {Lueks, Wouter and Mokbel, Bassam and Biehl, Michael and Hammer, Barbara},
+  keywords = {Machine Learning (cs.LG), Information Retrieval (cs.IR), FOS: Computer and information sciences, FOS: Computer and information sciences},
+  title = {How to Evaluate Dimensionality Reduction? - Improving the Co-ranking Matrix},
+  publisher = {arXiv},
+  year = {2011},
+  copyright = {arXiv.org perpetual, non-exclusive license}
+}
+@article{kraemer2018dimred,
+  doi = {10.32614/rj-2018-039},
+  url = {https://doi.org/10.32614/rj-2018-039},
+  year = {2018},
+  publisher = {The R Foundation},
+  volume = {10},
+  number = {1},
+  pages = {342},
+  author = {Guido Kraemer and Markus Reichstein and Miguel, D. Mahecha},
+  title = {{dimRed} and {coRanking} - Unifying Dimensionality Reduction in R},
+  journal = {The R Journal}
+}
+@article{chen2009local,
+  doi = {10.1198/jasa.2009.0111},
+  url = {https://doi.org/10.1198/jasa.2009.0111},
+  year = {2009},
+  month = mar,
+  publisher = {Informa {UK} Limited},
+  volume = {104},
+  number = {485},
+  pages = {209--219},
+  author = {Lisha Chen and Andreas Buja},
+  title = {Local Multidimensional Scaling for Nonlinear Dimension Reduction,  Graph Drawing,  and Proximity Analysis},
+  journal = {Journal of the American Statistical Association}
+}
+@article{lee2009quality,
+  doi = {10.1016/j.neucom.2008.12.017},
+  url = {https://doi.org/10.1016/j.neucom.2008.12.017},
+  year = {2009},
+  month = mar,
+  publisher = {Elsevier {BV}},
+  volume = {72},
+  number = {7-9},
+  pages = {1431--1443},
+  author = {John A. Lee and Michel Verleysen},
+  title = {Quality assessment of dimensionality reduction: Rank-based criteria},
+  journal = {Neurocomputing}
+}
+@article{venna2006local,
+  doi = {10.1016/j.neunet.2006.05.014},
+  url = {https://doi.org/10.1016/j.neunet.2006.05.014},
+  year = {2006},
+  month = jul,
+  publisher = {Elsevier {BV}},
+  volume = {19},
+  number = {6-7},
+  pages = {889--899},
+  author = {Jarkko Venna and Samuel Kaski},
+  title = {Local multidimensional scaling},
+  journal = {Neural Networks}
+}
