--- src/dimensionality_reduction/metrics/density/config.vsh.yaml
@@ -1,32 +1,0 @@
-__merge__: ../../api/comp_metric.yaml
-functionality:
-  name: "density"
-  namespace: "dimensionality_reduction/metrics"
-  description: "density preservation: correlation of local radius with the local radii in the original data space"
-  info:
-    v1_url: openproblems/tasks/dimensionality_reduction/metrics/density.py 
-    v1_commit: c2470ce02e6f196267cec1c554ba7ae389c0956a
-    metrics:
-      - id: density
-        label: Density
-        description: "density preservation: correlation of local radius with the local radii in the original data space"
-        min: 0
-        max: -1
-        maximize: true
-  resources:
-    - type: python_script
-      path: script.py
-platforms:
-  - type: docker
-    image: "python:3.10"
-    setup:
-      - type: python
-        packages:
-          - scipy
-          - numpy
-          - "anndata>=0.8"
-          - typing
-          - umap-learn
-  - type: nextflow
-    directives: 
-      label: [ lowmem, midcpu ]
