--- src/dimensionality_reduction/metrics/nn_ranking/config.vsh.yaml
@@ -1,30 +1,0 @@
-__merge__: ../../api/comp_metric.yaml
-functionality:
-  name: "nn_ranking"
-  namespace: "dimensionality_reduction/metrics"
-  description: A set of metrics from the pyDRMetrics package.
-  info:
-    v1_url: openproblems/tasks/dimensionality_reduction/metrics/nn_ranking.py 
-    v1_commit: a796e02e13a43e8861b124edbb9d287f162d4a14
-    metrics:
-      - id: nn_ranking
-        label: NN_Ranking
-        description: "A set of metrics from the pyDRMetrics package."
-  resources:
-    - type: python_script
-      path: script.py
-platforms:
-  - type: docker
-    image: "python:3.10"
-    setup:
-      - type: python
-        packages:
-          - numba
-          - numpy
-          - typing
-          - "anndata>=0.8"
-          - scipy
-          - scikit-learn
-  - type: nextflow
-    directives: 
-      label: [ lowmem, lowcpu ]
